address review comments

+++ numpy/lib/_function_base_impl.py
@@ -4586,7 +4586,7 @@ def _lerp(a, b, t, out=None):
     # asanyarray is a stop-gap until gh-13105
     lerp_interpolation = asanyarray(add(a, diff_b_a * t, out=out))
     subtract(b, diff_b_a * (1 - t), out=lerp_interpolation, where=t >= 0.5,
-             casting='unsafe')
+             casting='unsafe', dtype=lerp_interpolation.dtype)
     if lerp_interpolation.ndim == 0 and out is None:
         lerp_interpolation = lerp_interpolation[()]  # unpack 0d arrays
     return lerp_interpolation
